fix imports and move model
